Actually added powerup
for some reason gitkraken did not stage my changes. Anyway, the powerup is in but i couldn't find an elegant way to integrate it with the curret .csv template system without making it into another cell type. Also the powerups aren't displaying for some reason, i made sure to add them to the list of csv files in the boardaiagent